Update #presign for latest Shrine
Shrine 2.11+ allows #presign to return a simple Hash, so we can stop
using OpenStruct. Also, it's now recommended to include :method as well,
so that the client knows which HTTP verb it should it for the upload
request.

This change should make shrine-google_cloud_storage work with Uppy's
AwsS3 plugin.