Fix #clear! potentially not deleting past 2nd page
Google::Cloud::Storage::File::List#next doesn't appear to modify the
current list, it just returns a new list. Even if that wasn't the case,
that functionality might be removed in the future as it's not
documented.

So we switch to using the return value of #next, as is documented.